pyproject.toml
[project]
description = "DevOps AI Agent — Automated story-to-code pipeline"
name = "devops-ai-agent"
version = "0.1.0"
requires-python = ">=3.10"
license = {text = "MIT"}
readme = "README.md"

dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "gitpython>=3.1",
    "watchdog>=4.0",
]

[project.optional-dependencies]
zendesk = ["zenpy>=2.0"]
webhook = ["flask>=3.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.24", "ruff>=0.8"]

[project.scripts]
dai = "src.cli:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
